:recycle: refactor: S3 관련 로직 분리 (#28)
